requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
base58==2.1.1
bitarray==3.8.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.1
cryptography==46.0.5
cytoolz==1.1.0
distro==1.9.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==6.0.0b1
filetype==1.2.0
Flask==3.1.3
frozenlist==1.8.0
google-auth==2.48.0
google-genai==1.64.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-core==1.2.14
langchain-google-genai==4.2.1
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.8
langsmith==0.7.6
MarkupSafe==3.0.3
multidict==6.7.1
nest-asyncio==1.6.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
paho-mqtt==2.1.0
parsimonious==0.10.0
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
pyunormalize==17.0.0
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rlp==4.1.0
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.4
toolz==1.1.0
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
web3==7.14.1
websockets==15.0.1
Werkzeug==3.1.6
x402==2.2.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
zyndai-agent==0.2.3
